Fix insders build and make CI notification pretty (#814)
I don't know why but for some reason we can't import
from`'@sourcegraph/cody-shared'` directly. I've seen this error a couple
of times now. This time it was only noticable on CI though (which is
weird, I thought we fixed that but perhaps the PR was not based on the
latest changes on `main` yet).

Anyway, this should fix the import and also makes the CI notification
more useful.

## Test plan

```
α vscode (ps/fix-insiders)
pnpm run build

> cody-ai@0.8.0 build /Users/philipp/dev/cody/vscode
> tsc --build && pnpm run -s _build:esbuild:desktop && pnpm run -s _build:esbuild:web && pnpm run -s _build:webviews --mode production


  dist/extension.node.js      6.0mb ⚠️
  dist/extension.node.js.map  9.7mb

⚡ Done in 228ms

  dist/extension.web.js      5.0mb ⚠️
  dist/extension.web.js.map  7.7mb

⚡ Done in 167ms
vite v4.4.3 building for production...
✓ 438 modules transformed.
../dist/webviews/index.html                0.78 kB
../dist/webviews/codicon-2d58c995.ttf     70.78 kB
../dist/webviews/index-948936fc.css       58.00 kB
../dist/webviews/index.js              1,759.47 kB │ map: 12,839.81 kB
✓ built in 2.20s
```

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->