Autocomplete: Surface rate limit and other errors (#851)
Closes #634

This PR adds a UI indicator for autocomplete errors. We special case
rate limiting errors to provider more information.

- When a rate limit error is encountered, we show it only once per rate
limit interval (the UI indicator will reset with a VS Code restart).
When the option is selected in quick pick, we open
https://about.sourcegraph.com/blog/increasing-the-completions-rate-limit
(the only resource I found that we have for rate limits)
- When any other (e.g. a network error) is encountered, we show the
error message and open the output channel. We also add a new `error` API
similar to the `debug` one but that will always log errors to the output
channel.

## ToDo

- [x] Add a new `error` API similar to the `debug` one but that will
always log errors to the output channel.
- [x] Add test cases

## Test plan


https://github.com/sourcegraph/cody/assets/458591/efa5f660-706e-47db-96e1-63b5f7b2d2fc



<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Tim Lucas <t@toolmantim.com>