VS Code: use singleton storage providers (#876)
- This PR migrates `localStorage` and `secretStorage` to a singleton
pattern. Instead of initializing these entities upon the extension
activation, they are initialized in the module scope and can be imported
anywhere in the VS Code extension code. This removes a lot of
boilerplate required for prop-drilling these entities from the extension
init function to consumers. See the example of the boilerplate removal
here:
https://github.com/sourcegraph/cody/commit/f445839ee39a44b11fdedcd2be0c0c883357edde.
- Touched on similar pain points in [the PR
discussion](https://github.com/sourcegraph/cody/pull/636#discussion_r1298003228)
- Part of #874
- Planned follow-up: migrate other `localStorage` and `secretStorage`
consumers to importing singletons instead of using them as arguments.
Same as done
[here](https://github.com/sourcegraph/cody/commit/f445839ee39a44b11fdedcd2be0c0c883357edde).

The end goal of this effort (as stated in
https://github.com/sourcegraph/cody/issues/874) is to decouple
`getConfiguration` from the other entities in the codebase and introduce
atomic and reactive getting for config values.