E2E tests: Upload recording artifacts and fix flaky test (#919)
This fixes the flaky E2E test. It tried to bring the Cody sidebar into
the foreground but it was never hidden in the first place. I also added
additional assertions about the `Hola` case that was not asserted on
before.

This PR also brings back video uploads from failed cases so we can
easily debug CI errors again.

## Test plan

- Checked with a failing E2E test
<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->