Autocomplete: Tweak StarCoder temperature (#966)
This changes the default temperature for StarCoder from `0.4` to `0.2`
which is also used in [their own
testing](https://github.com/bigcode-project/bigcode-evaluation-harness).
In my experiments, this made the model behave a bit more predictable
while still providing enough variation.

**CONTROVERSIAL:** I also tweaked the example Bee added in #945. I’m
happy to revert though. StarCoder was not completing this and I think it
was because there are two reference function in that file that are
implemented the same way with the same line missing so starcoder did
pattern match and thought that this is the expected result. I feel that
the example is a bit contrived so I changed it into something that I
would encounter while developing.

## Test plan

Ran the completions test suite and made sure the output looked good:

- `cd vscode`
- `pnpm run generate:completions`

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->