Autocomplete: Add bare bones statistics logging and UI (#973)
I often find myself wanting to know when exactly we count something as
suggested vs. when we count it as accepted. Looking at the output tab is
not super helpful here either so I added a small UI to the trace view.

This also adds the "infra" to support the trace view to show other views
that re-render which I want to use for the section observer shortly.

I’m specifically hoping that this will help with #923.

## Test plan


https://github.com/sourcegraph/cody/assets/458591/2279928c-4056-4d0e-9f56-d5bdff6dbc03


<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->