Agent: add notification that clears the last autocomplete candidate (#998)
Clearing the last candidate ensures that autocomplete id's and cached
potential suggestions are not reused. This adds a notification an
extension can send when it no longer wishes the consider the last
candidate. The most common reason is that the suggestion was accepted.
## Test plan
Tested using JetBrains extension 
Accepted an autocomplete and then immediately removed it and triggered
another
Got the same suggestion but verified the id was different.

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->