Intelligent Cody: Supercharge hover text (#983)
For the completions path, we want to be able to fill in the _meta_
details about symbols (shapes of interfaces, classes, etc) and not
necessarily the full implementation. This context seems to be enough to
give a boost to the completion engine.

This PR updates the completions code path to return the _hover text_
which provides sufficient information for Go and TypeScript (and
possibly other languages). We go further in our approach by not only
selecting the hover text of the symbol itself (which should be
equivalent to the *definition*'s hover text), but also the hover text of
associated type definitions and implementations.

## Test plan

Updated unit tests, manual testing.


![image](https://github.com/sourcegraph/cody/assets/103087/c16968ba-bff2-4fef-b945-d45ba5e07bc5)

---------

Co-authored-by: Philipp Spiess <hello@philippspiess.com>