update claude infill prompt to fix indent issue (#990)
RE: https://sourcegraph.slack.com/archives/C05AGQYD528/p1694095905077429

Fix the issue mentioned in the attached slack thread where
autocompletion for multi-line is not activated in an empty line.

This was caused by the modified indentation for the active cursor lines
that we currently use with the prompt that does not support "infill".

To get the infill to work with the updated prompt, the indentation needs
to be reserved.

This PR added a new tail and head in a format that works with the new
prompt in order to workaround the issue. Also attached test result to
the test suite for record.

### Before:

Multi-line completions are generated after a opening bracket

### After

Can generate multi-line completions on new empty line as expected in our
test suite.

## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

Run test suite to compare results between current claude instant and
claude instant supporting infill.

Set: `"cody.autocomplete.advanced.model": "claude-instant-infill"`


![image](https://github.com/sourcegraph/cody/assets/68532117/0d290162-68d6-413f-a99c-b8dcf564e8aa)

P.S.: this is just a workaround to get Claude to try to behave as a code
completion ai. It is not expected to work as well as llms that are
specifically trained for this purpose. However the improvement made by
the prompt change is still noticeable