Remove remove unstable-azure provider (#1003)
Now that the 5.1.8 release of Sourcegraph has experimental support for
`azure-openai` as an autocomplete provider, this unstable provider that
connects directly to azure is no longer needed. To use azure for
autocomplete a user should choose the `unstable-openai` provider and
connect to a sg instance that is configured to use azure's openai apis.