JetBrains: correctly manage chat transcript (#56671)
Previously, the state of the chat transcript went out of sync between
what was visible in the JetBrains UI and what the Cody worker (aka. agent)
modeled internally. This PR fixes the problem by sending a
`transcript/reset` notification when the user triggers the "Reset
conversation with Cody" action.

This PR must be tested alongside the PR https://github.com/sourcegraph/cody/pull/1073

Former-commit-id: dffe1685f034f9dcac2cb8bb836c808634b50479