Autocomplete: Only persist last candidate if it would be shown in the UI (#1059)
Part of #810 

As part of the audit for #810 I noticed a potential here in when we save
the last candidate:

- Whenever the request manager yields a result (which might be for a
completion that was already cancelled) we currently persist the returned
result as the last candidate.
- For already cancelled completions, this could mean that the completion
is never being shown on the screen, yet it will set the `lastCandidate`
and can later be picked up as the last candidate immediatly
- This also means that if the document changed during a completion
request (which would cancel the completion) we could have "borked" cases
in the lastCandidate that used the `document.getText` APIs at various
states of the document and thus this could be an explaination of the
issues we're seeing in #810
- Conceptually, I think this logic makes a lot of sense: If we never
render ghost text, we should not cache something as the last candidate

One thing to note here is the interop with the
`completeSuggestWidgetSelection` flag: When it is set, I expect this to
fully work. However, when it is not set, we might be causing a few
unnecessary requests. Consider this case:


https://github.com/sourcegraph/cody/assets/458591/6ac8caa6-1380-4fd2-b9ba-26d0c250a6db

Every time I navigate between the suggestions, we trigger a new
autocomplete request yet nothing in the prefix/suffix changes. Since the
completions aren't ever visible (they would only show up if the
selection happens to overlap with the completion), we never save a last
candidate and thus every subsequent navigation causes a new request.

This is currently not a problem as we gave the network request cache but
we want to remove that eventually. As the completion would not ever be
visible for this case, this does not make a difference for the user
though.

## Test plan

- Play around with autocomplete and observe no obvious issues
- Tests still pass
- Unfortunalty we lack a consistent repro for #810 

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->