Autocomplete: Use the same context window size for StarCoder (#1058)
I noticed an issue in the StarCoder setup that might impact our A/B
tests: We currently define a lower context window size than we do for
Claude Instant. Since StarCoder supports up to 8k token windows (a lot
more than we need), I’m going to raise it to the same limits we have for
Claude Instant (~2k). This should ensure that we include the same
context in both branches.

I plan on cherry picking this onto the stable branch and pushing this as
a patch release so we have fairer conditions in the A/B test.

## Test plan

- Enable StarCoder by being connected to dotcom, setting the provider to
Fireworks, and setting the model to StarCoder 16b
- Open a lot of typescript files and trigger a completion in a very long
file
- Observer that the prompt now contains up to 8k characters 

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Valery Bugakov <skymk1@gmail.com>