Autocomplete: Fix support for documents with \r\n line endings (#1124)
Fixes #1092

We currently have issues with files that contain CRLF line endings. This
PR fixes a few of the most common gotchas and adds some additional test.
We likely need to conduct further testing though, I saw that there are
still a few cases of `split('\n')` remaining.

Since this change normalizes the prefix and suffx in the document
context to not contain `\r\n` anymore, I lot of the downstream errors
should already be fixed I believe. One potential for issue is if the LLM
decides to return `\r\n` terminated line endings now which might still
happen I suppose. Also context snippets might still contain `\r\n` when
passed to the prompt (we do not have logic to truncate stuff based on
this though, so I don't anticipate this to cause issues).

I do not have access to a Windows machine right now to verify though. 😬 

## Test plan

- Added tests

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->