feat: New "Save Code to File" button and other chat code block style updates (#1119)
Updates to code blocks rendered in chat:
* Moves the code block buttons to the bottom of the code blocks and
always shown (instead of just on hover)
* Adds a `Save Cody to File…` button
* For dark theme users, updates the chat code block background to match
the editor background

## Details

This commit adds a new insertButtonOnSubmit prop to the Chat and
CodeBlocks components. This allows passing in a callback function that
will be triggered when the insert button is clicked.

The CopyButtonProps interface is updated to have separate callbacks for
copy and insert actions. The CodeBlocks component is updated to
conditionally render insert and insert new file buttons if the callback
props are passed in.

This enables IDE integrations to handle insert actions separately from
copy actions.

When `Save to File` button is pressed, user will be prompted to select a
new file file to insert the code to. If they selected an existing file,
the content will be replaced. See demo video below for details.

In addition, I've also updated the code block action styles to match
Tim's design on
[figma](https://www.figma.com/file/9112BsKsJc1BpO2j8XYwLL/%F0%9F%A4%96-Cody-VS-Code-%5BMain%5D?type=design&node-id=5231-8881&mode=design&t=nrLPZluUa8zxyO6f-4).

## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

- start Cody from this branch
- click on "Save to File" button below the code block



https://github.com/sourcegraph/cody/assets/68532117/3e86f8be-c9d7-4f1d-b634-eb4849eaab12

---------

Co-authored-by: Tim Lucas <t@toolmantim.com>