Agent: add support to cancel chat (#1162)
Previously, sending `$/cancelRequest` notification for `recipes/execute`
did not correctly cancel the chat request. This PR fixes the problem
with two changes:

- Don't return from `recipes/execute` until the last
`chat/updateMessageInProgress` has been sent. This was necessary for
cancelation to work correctly because the request needs to be running
for it to be cancelable. We delete the `CancellationTokenSource` after a
request returns.
- Correctly propagate the cancelation token from the agent into the chat
client.


## Test plan

See updated `index.test.ts`

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->