Autocomplete: Add logging for partial acceptances (#1214)
This PR adds logging for partially accepted completions #1204.

A completion will be logged as partially accepted, when at least one
full word of the completion is forward typed (either by manually typing
or using the cmd + arrow right shortcut)

After a completion is partially accepted, the event will fire for every
additional keystroke. This is so that in analytics we can use the
`max(acceptedLength)`. If we only count partially accepted completions,
we thus have to add a `distinct id` check (cc @chenkc805 @kelsey-brown)

## Test plan


https://github.com/sourcegraph/cody/assets/458591/378bf18a-3ff6-406c-8111-bdd77be97ecd



<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->