BFG: add new graph context engine
Previously, the "Graph Context" feature flag was a boolean that used an
LSP-powered solution when it had the value `true`. This PR changes the
"Graph Context" feature flag into an enum with the values: `lsp`, `bfg`
and `none` (default). The `bfg` engine is implemented by a new JSON-RPC
server that speaks a bespoke BFG-protocol. BFG is implemented in the
sourcegraph/sourcegraph repository and distributed via binary releases
in the sourcegraph/bfg repository (https://github.com/sourcegraph/bfg/releases).

The design of BFG is documented in RFC 829
https://docs.google.com/document/d/1smI8mdRhJoZ23uGYh4bxl2iDd-b1kp3qMNIOKwXwacg/edit#heading=h.trqab8y0kufp