Autocomplete: Add bookkeeping to reuse completion IDs over multiple suggestions (#1230)
This PR will add a new book keeping for completion IDs. These are
different from the existing suggestion IDs in that:

- Suggestion IDs are used for one suggestion life cycle: One time a
completion is rendered on the screen
- Completion IDs are used to identify a specific completion that is
printed in a specific document location

The new change makes it so that we will log multiple completions events
with the same id. This will help us identify how useful the current
cache behavior is.

## Test plan


https://github.com/sourcegraph/cody/assets/458591/5e6604f8-e4a3-4b38-ab5c-16776677d265



<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Dominic Cooney <dominic.cooney@sourcegraph.com>