Autocomplete: add single completion trigger queries (#1259)
- Trigger single-line completions on `type`/`interface` declarations in
typescript and `struct` declarations in go. Powered by tree-sitter
queries.
- If a parse tree is available in `detectMultiline,` we query the
document for node types mentioned above.
- If singleline triggers are found under the cursor, we exit early. The
current logic applies for all the other cases (e.g., we return
`blockStart` as a multiline trigger if the document prefix ends with
it).
    - How do we understand if the node found is under the cursor?
        - Queries captures `block_start` nodes — `{` characters. 
- Then, we ensure that the `block_start` position equals the cursor
position. If that's not the case, we discard found singleline completion
trigger.
- Potentially, we can find the `block_start` node imperatively from a
`trigger` node using parse-tree operations, but I'm not sure if it will
be easy to keep the same logic across languages.
- Adds a new user `cody.autocomplete.experimental.syntacticTriggers`
setting to enable the new functionality.
- Adds snapshot tests for the new tree-sitter queries.