JetBrains: handle nullable agent properties including `displayText`
Previously, chat was not working when using the latest build of Cody
because the `ChatMessage.displayText` had now become optional even if it
had a `@NotNull` annotation (causing the plugin to crash). This PR
updates the `ChatMessage` class to allow `displayText` to be nullable,
falling back to `.text` or `""`. To prevent a similar surprise from
happening other places, this PR also adds `@Nullable` annotations on
other agent protocol messages and adds appropriate error handling.


Former-commit-id: 820be69a99dd061b0393907ded7662fe249ca63f