add: line comment detection to completion latency (#1316)
This adds detection of line comments to the getLatency function to
handle comment lines differently. The base latency is increased to
1000ms for comment lines.

Also added markdown, plaintext, XML to the low-performance language
list. (feel free to remove if you have opinions on this).

NOTE: All changes in this PR is hidden behind the minLatency feature
flag

TODO: put it behind a different feature flag for comments, but want to
check with the team and see if this makes sense first.

## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

Test cases were added for the new isLineComment function.