Use vscode-uri for the mock/shim Uri (#1335)
This uses a proxy over `vscode-uri`'s `URI` class, with a fallback to
the `Utils` class as a replacement for `Uri`. This should be compatible
and avoids copying the `Uri` implementation from VS Code.

It's not possible to use the proxy as a type (so you must use
`vscode_types.Uri` for files that aren't importing `vscode`), but you
_can_ use it in `instanceof`, and TS also understands that it has all
the methods from the standard `vscode_types.Uri` class.

## Test plan

Tests with `pnpm test:unit`. Includes new tests to ensure static
methods, instance members and `instanceof`s work as expected.