Agent: downgrade to Node.js v16 for standalone binaries
Previously, the standalone binaries for the agent embedded the latest
version of Node.js, which at this time is most likely v20. Some users
are unable to the the JetBrains plugin because they are hitting on the
following network error in the agent

```
routines:final_renegotiate:unsafe legacy renegotiation disabled
```

This article here suggests that the error only happens in Node.js v17
and newer.
https://jfrog.com/help/r/artifactory-why-am-i-observing-unsafe-legacy-renegotiation-disabled-during-npm-install/artifactory-why-am-i-observing-unsafe-legacy-renegotiation-disabled-during-npm-install
This PR is an attempt to fix this problem by using Node.js v16 in the
agent.