JetBrains: Agent cleanup on project close (#11)
The agent was not getting shutdown because the project was already
disposed when stop was called which bypassed the process.

This shutdowns the agent process when the project scoped service for
agent is disposed, and removes the project listener that had been
responsible for shutting down the agent but was calling it the project
was disposed.

closes https://github.com/sourcegraph/jetbrains/issues/8
## Test plan
`./gradlew  :runIDE`  
open a project
Open activity manager verify 1 cody-agent
Open a second project in a new window - verify activity manager shows 2
cody-agent
Test chat & autocomplete in each window
Close one project - verify activity manager shows 1 cody-agent
Test chat & autocomplete in remaining window
Exit current Editor complete
verify activity manager shows no cody-agent



<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles

Why does it matter? 

These test plans are there to demonstrate that are following industry
standards which are important or critical for our customers.
They might be read by customers or an auditor. There are meant be simple
and easy to read. Simply explain what you did to ensure
your changes are correct!

Here are a non exhaustive list of test plan examples to help you:

- Making changes on a given feature or component: 
- "Covered by existing tests" or "CI" for the shortest possible plan if
there is zero ambiguity
  - "Added new tests" 
- "Manually tested" (if non trivial, share some output, logs, or
screenshot)
- Updating docs: 
  - "previewed locally" 
  - share a screenshot if you want to be thorough
- Updating deps, that would typically fail immediately in CI if
incorrect
  - "CI" 
  - "locally tested" 
-->