Autocomplete: Implement client-side timeouts (#1355)
This PR adds client side timeouts as discussed in [this Slack
thread](https://sourcegraph.slack.com/archives/C05497E9MDW/p1696521645079319).

The reason for this are the following:

Our current server-side backend does request retrying for a fixed number
of times. For Autocomplete however, latency is of upmost importance so
we don't really want to retry requests for a long time. In some cases,
we see requests being retried for up to 30 minutes which is just putting
additional pressure on the downstream processes without providing a
great UX. Since timeouts are highly client specific, we decided that
adding those to the client is the best approach for now. All of our
middleware respect aborted requests so all we need to do is to start a
race between a timeout and abort the request if that is hit.

**Note:** This PR is best reviewed [with whitespace changes
ignored](https://github.com/sourcegraph/cody/pull/1355/files?w=1).

## Test plan

- Modify the code to have a timeout below 100ms
- Trigger an autocomplete request
- Observe that a timeout error is thrown

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->