feat: prevent unhelpful suggestion at empty file end line (#1330)
change: prevent completions at empty file end

Proposing to not trigger completions when the cursor is at the start of
the last line of a file AND the line above is empty.

This avoids unwanted completions when finishing up a file, or when it's
clear that we do not have enough context to make a helpful suggestion.

### Proposal

Going through this
[thread](https://sourcegraph.slack.com/archives/C05AGQYD528/p1695856420061639),
I noticed we provide a lot of suggestions for new line, but based on the
number a lot of them are not very unhelpful and probably can be cut
down. One thing that is most obvious to me is the suggestions i've been
getting at the end of file.

Most of the time the suggestions from there are just PURE guesses, here
are some examples:


![image](https://github.com/sourcegraph/cody/assets/68532117/fa37dce4-ab7e-468a-96c8-21a82a6f2a4d)


![image](https://github.com/sourcegraph/cody/assets/68532117/d46ad646-7cce-4445-af82-a7bc1e029b1d)


![image](https://github.com/sourcegraph/cody/assets/68532117/41ceebc6-c7be-452a-bfde-18e2401f28a0)


This is why I propose not to trigger completions automatically at this
specific position.

However, this only applies to end lines where the current cursor is at
position 0. if there is an indentation at the start, or if this is not
an empty line, then we know this is most likely a WIP, and therefore a
signal to let us know that this is not the end of the file and we might
have context to continue (even though it might not be true, and that's
why we still want to trigger suggestion automatically here):

![image](https://github.com/sourcegraph/cody/assets/68532117/5e797648-d498-44ad-98ac-c7ce7449410c)

We also don't want to prevent suggestions from showing up if the line
above is not empty, as that could potentially be source of context, e.g.
a comment:

![image](https://github.com/sourcegraph/cody/assets/68532117/b60c8cdf-f698-4e84-9605-4d8605ffafe6)

Thoughts?

## Test plan

<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

Move your cursor to the end of file.

Also added unit tests

---------

Co-authored-by: Philipp Spiess <hello@philippspiess.com>