Autocomplete: LSP light graph context (#1326)
There are some bugs yet which prevents the `Banan` from being inlined
but this is a WIP PR for some changes that should bring down the LSP
usage by a _ton_.

1. We no longer rely on document sections to match context. Even with
improvements in #1314 this would easily explode to too many symbols.
2. We still leverage cursor navigation based preloading to reduce
latencies.
3. Instead of a range, we use the current line, previous line for
context only.
4. I also want to _recurse_ into the current line. My idea is to fix for
the banana example that Olaf came up with.
5. The actual LSP calls were also changed. We now call `hover` on
everything that looks like a symbol. We recurse the current line. this
means for every symbol in this line we also goToDef and run generate the
hover tooltip for the first 50 symbols inside the range of the
definition of that symbol. That is enough to cover the bananas example.
6. Furthermore this now also awaits the retrieval promise during the
autocomplete request. This means that even when the preloading is not
finished yet/or you just typed a new character, we will wait for the
context to be loaded.
7. Preloading is cached per line including the line text. That means
that every change to the line will evict the cache so that we always
prioritize the latest data for context.

## Test plan

Test with the infamous bananas example. The timing on the right is with
0 caches so for this example it fits the 100ms context budget but we
might need to add some further limits for larger implementations

<img width="1362" alt="Screenshot 2023-10-09 at 16 04 13"
src="https://github.com/sourcegraph/cody/assets/458591/6d34624f-55f0-4fd7-b0b7-03ebed2f35c1">


<!-- Required. See
https://docs.sourcegraph.com/dev/background-information/testing_principles.
-->

---------

Co-authored-by: Dominic Cooney <dominic.cooney@sourcegraph.com>